refactor: Splt HistoryCleanupMapper for common method and process related method
