+++ db/rdbms/src/main/java/io/camunda/db/rdbms/sql/ProcessBasedHistoryCleanupMapper.java
@@ -10,11 +10,9 @@ package io.camunda.db.rdbms.sql;
 import io.camunda.util.ObjectBuilder;
 import java.time.OffsetDateTime;
 
-public interface ProcessBasedHistoryCleanupMapper {
+public interface ProcessBasedHistoryCleanupMapper extends HistoryCleanupMapper {
 
   int updateHistoryCleanupDate(UpdateHistoryCleanupDateDto dto);
-
-  int cleanupHistory(CleanupHistoryDto dto);
 
   record UpdateHistoryCleanupDateDto(long processInstanceKey, OffsetDateTime historyCleanupDate) {
 
@@ -36,36 +34,6 @@ public interface ProcessBasedHistoryCleanupMapper {
       @Override
       public UpdateHistoryCleanupDateDto build() {
         return new UpdateHistoryCleanupDateDto(processInstanceKey, historyCleanupDate);
-      }
-    }
-  }
-
-  record CleanupHistoryDto(int partitionId, OffsetDateTime cleanupDate, int limit) {
-
-    public static class Builder implements ObjectBuilder<CleanupHistoryDto> {
-
-      private int partitionId;
-      private OffsetDateTime cleanupDate;
-      private int limit;
-
-      public Builder partitionId(final int partitionId) {
-        this.partitionId = partitionId;
-        return this;
-      }
-
-      public Builder cleanupDate(final OffsetDateTime cleanupDate) {
-        this.cleanupDate = cleanupDate;
-        return this;
-      }
-
-      public Builder limit(final int limit) {
-        this.limit = limit;
-        return this;
-      }
-
-      @Override
-      public CleanupHistoryDto build() {
-        return new CleanupHistoryDto(partitionId, cleanupDate, limit);
       }
     }
   }
